requirements.txt
requests==2.23.0
python-dotenv==0.10.3
Whoosh==2.7.4
praw==7.1.0
